+++ src/search/merge_and_shrink/merge_scoring_function_dfp.h
@@ -13,7 +13,7 @@ protected:
 public:
     MergeScoringFunctionDFP() = default;
     virtual ~MergeScoringFunctionDFP() override = default;
-    virtual std::vector<int> compute_scores(
+    virtual std::vector<double> compute_scores(
         FactoredTransitionSystem &fts,
         const std::vector<std::pair<int, int>> &merge_candidates) override;
 };
